fix(#51): recognizedHours Float 형으로 변경

+++ finda-volunteer/src/main/kotlin/finda/findavolunteer/domain/volunteer/data/VolunteerDocumentData.kt
@@ -23,5 +23,5 @@ data class ParticipantInfo(
     val grade: Int,
     val classNum: Int,
     val num: Int,
-    val recognizedHours: Int
+    val recognizedHours: Float,
 )
